+++ packages/ember-application/lib/system/engine.js
@@ -17,9 +17,9 @@ import EngineInstance from './engine-instance';
 import isEnabled from 'ember-metal/features';
 import symbol from 'ember-metal/symbol';
 import Controller from 'ember-runtime/controllers/controller';
-import TextField from 'ember-views/views/text_field';
-import TextArea from 'ember-views/views/text_area';
-import Checkbox from 'ember-views/views/checkbox';
+import TextField from 'ember-htmlbars/components/text_field';
+import TextArea from 'ember-htmlbars/components/text_area';
+import Checkbox from 'ember-htmlbars/components/checkbox';
 import LinkToComponent from 'ember-routing-views/components/link-to';
 import RoutingService from 'ember-routing/services/routing';
 import ContainerDebugAdapter from 'ember-extension-support/container_debug_adapter';
