+++ packages/ember-application/tests/system/visit_test.js
@@ -7,7 +7,7 @@ import ApplicationInstance from 'ember-application/system/application-instance';
 import Route from 'ember-routing/system/route';
 import Router from 'ember-routing/system/router';
 import View from 'ember-views/views/view';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 import compile from 'ember-template-compiler/system/compile';
 import jQuery from 'ember-views/system/jquery';
