+++ packages/ember-glimmer/tests/integration/components/fragment-components-test.js
@@ -1,7 +1,7 @@
 import { moduleFor, RenderingTest } from '../../utils/test-case';
 import { strip } from '../../utils/abstract-test-case';
 import { set } from 'ember-metal/property_set';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 
 moduleFor('Components test: fragment components', class extends RenderingTest {
   getCustomDispatcherEvents() {
