+++ packages/ember-glimmer/tests/utils/helpers.js
@@ -1,5 +1,8 @@
 export { default as Helper, helper } from 'ember-glimmer/helper';
-export { default as Component } from 'ember-glimmer/ember-views/component';
+export { default as Component } from 'ember-glimmer/component';
+export { default as Checkbox } from 'ember-glimmer/components/checkbox';
+export { default as TextArea } from 'ember-glimmer/components/text_area';
+export { default as TextField } from 'ember-glimmer/components/text_field';
 export { DOMHelper } from 'glimmer-runtime';
 export { InteractiveRenderer, InertRenderer } from 'ember-glimmer/renderer';
 export { default as compile } from 'ember-glimmer/ember-template-compiler/system/compile';
