+++ packages/ember-htmlbars/lib/component.js
