+++ packages/ember-htmlbars/lib/components/text_area.js
@@ -2,7 +2,7 @@
 @module ember
 @submodule ember-views
 */
-import Component from 'ember-views/components/component';
+import Component from '../component';
 import TextSupport from 'ember-views/mixins/text_support';
 
 /**
