+++ packages/ember-htmlbars/tests/helpers/input_test.js
@@ -4,8 +4,8 @@ import View from 'ember-views/views/view';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
 import compile from 'ember-template-compiler/system/compile';
 import ComponentLookup from 'ember-views/component_lookup';
-import TextField from 'ember-views/views/text_field';
-import Checkbox from 'ember-views/views/checkbox';
+import TextField from 'ember-htmlbars/components/text_field';
+import Checkbox from 'ember-htmlbars/components/checkbox';
 import EventDispatcher from 'ember-views/system/event_dispatcher';
 import buildOwner from 'container/tests/test-helpers/build-owner';
 import { OWNER } from 'container/owner';
