+++ packages/ember-htmlbars/tests/integration/component_lifecycle_test.js
@@ -1,7 +1,7 @@
 import jQuery from 'ember-views/system/jquery';
 import compile from 'ember-template-compiler/system/compile';
 import ComponentLookup from 'ember-views/component_lookup';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 import { runAppend, runDestroy } from 'ember-runtime/tests/utils';
 import run from 'ember-metal/run_loop';
 import EmberView from 'ember-views/views/view';
