+++ packages/ember-htmlbars/tests/utils/helpers.js
@@ -1,5 +1,8 @@
 export { default as Helper, helper } from 'ember-htmlbars/helper';
 export { default as DOMHelper } from 'ember-htmlbars/system/dom-helper';
-export { default as Component } from 'ember-views/components/component';
+export { default as Component } from 'ember-htmlbars/component';
+export { default as Checkbox } from 'ember-htmlbars/components/checkbox';
+export { default as TextArea } from 'ember-htmlbars/components/text_area';
+export { default as TextField } from 'ember-htmlbars/components/text_field';
 export { InteractiveRenderer, InertRenderer } from 'ember-htmlbars/renderer';
 export { default as compile } from 'ember-template-compiler/system/compile';
