+++ packages/ember-testing/tests/helpers_test.js
@@ -4,7 +4,7 @@ import run from 'ember-metal/run_loop';
 import EmberObject from 'ember-runtime/system/object';
 import RSVP from 'ember-runtime/ext/rsvp';
 import EmberView from 'ember-views/views/view';
-import Checkbox from 'ember-views/views/checkbox';
+import Checkbox from 'ember-htmlbars/components/checkbox';
 import jQuery from 'ember-views/system/jquery';
 
 import Test from 'ember-testing/test';
