+++ packages/ember-views/tests/views/component_test.js
@@ -6,7 +6,7 @@ import inject from 'ember-runtime/inject';
 import { get } from 'ember-metal/property_get';
 
 import EmberView from 'ember-views/views/view';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 
 import { MUTABLE_CELL } from 'ember-views/compat/attrs-proxy';
 import buildOwner from 'container/tests/test-helpers/build-owner';
