+++ packages/ember-views/tests/views/text_area_test.js
@@ -1,6 +1,6 @@
 import EmberObject from 'ember-runtime/system/object';
 import run from 'ember-metal/run_loop';
-import TextArea from 'ember-views/views/text_area';
+import TextArea from 'ember-htmlbars/components/text_area';
 import { get } from 'ember-metal/property_get';
 import { set as o_set } from 'ember-metal/property_set';
