+++ packages/ember/tests/application_lifecycle_test.js
@@ -1,7 +1,7 @@
 import Application from 'ember-application/system/application';
 import Route from 'ember-routing/system/route';
 import run from 'ember-metal/run_loop';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 import jQuery from 'ember-views/system/jquery';
 import { compile } from 'ember-template-compiler';
 import { getTemplates, setTemplates } from 'ember-htmlbars/template_registry';
