+++ packages/ember/tests/controller_test.js
@@ -4,7 +4,7 @@ import run from 'ember-metal/run_loop';
 import { compile } from 'ember-template-compiler';
 import Application from 'ember-application/system/application';
 import EmberView from 'ember-views/views/view';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 import jQuery from 'ember-views/system/jquery';
 import { setTemplates, set as setTemplate } from 'ember-htmlbars/template_registry';
