+++ packages/ember/tests/default_initializers_test.js
@@ -1,6 +1,6 @@
 import Application from 'ember-application/system/application';
-import TextField from 'ember-views/views/text_field';
-import Checkbox from 'ember-views/views/checkbox';
+import TextField from 'ember-htmlbars/components/text_field';
+import Checkbox from 'ember-htmlbars/components/checkbox';
 
 import run from 'ember-metal/run_loop';
