+++ packages/ember/tests/helpers/link_to_test.js
@@ -7,7 +7,7 @@ import run from 'ember-metal/run_loop';
 import isEnabled from 'ember-metal/features';
 import alias from 'ember-metal/alias';
 import Application from 'ember-application/system/application';
-import Component from 'ember-views/components/component';
+import Component from 'ember-htmlbars/component';
 import ComponentLookup from 'ember-views/component_lookup';
 import jQuery from 'ember-views/system/jquery';
 import EmberObject from 'ember-runtime/system/object';
